+++ smsfee/report/report_unpaid_fee_bills_2folded.py
@@ -14,18 +14,16 @@ import babel
 logger = netsvc.Logger()
 result_acc=[]
      
-"""This is the main parsor that prints challans for academics and transport with 1 student per page 
-   class wise, other parsers that prints class wise challans, should be rmeoved """
-
-class unpaid_fee_challan_parser(report_sxw.rml_parse):
-    #this will be the only challans parsser called for
-    # acadimc fee, transport fee and other 
-    # currently called for clasess wside fees for trasport and academics
-    # later on will be adjusted to call the same parser for sing student challan prrintg both in academcis and transport 
+class report_unpaid_fee_bills_2folded(report_sxw.rml_parse):
+    
+    """
+    This is the paser called when 2 folded fee challans are to be printed, tested by shahid 29 DEC 17
+    
+    """
  
     def __init__(self, cr, uid, name, context):
  
-        super(unpaid_fee_challan_parser, self).__init__(cr, uid, name, context)
+        super(report_unpaid_fee_bills_2folded, self).__init__(cr, uid, name, context)
         self.result_temp=[]
         self.localcontext.update( {
             'get_today':self.get_today,
@@ -46,47 +44,28 @@ class unpaid_fee_challan_parser(report_sxw.rml_parse):
             'get_challan_header_linethree':self.get_challan_header_linethree,
             'get_challan_footer_one':self.get_challan_footer_one,
             'get_challan_footer_two':self.get_challan_footer_two,
-            'get_department_logo':self.get_department_logo,
          })
         self.context = context
 
     def get_challan_logo(self):
-        #this will return instute logo on challan
-        #another same method will return transport or academics logo, will be added soon
-        logo = 'No Logo'
-        company_id = self.pool.get('res.company').search(self.cr, self.uid,[])
-        #-------------Handling Only one Company is There are multiple companies blank space will be returned----------------        
-        if len(company_id)>1:
-            return ''
-        elif len(company_id) == 1:
-            company_recs = self.pool.get('res.company').browse(self.cr, self.uid, company_id)[0]
-            logo = company_recs.logo
-        return logo
-
-    def get_department_logo(self):
         rescompany_id = self.pool.get('res.company').search(self.cr, self.uid,[])
         #-------------Handling Only one Company is There are multiple companies blank space will be returned----------------        
         if len(rescompany_id)>1:
             return ''
         company_recs = self.pool.get('res.company').browse(self.cr, self.uid, rescompany_id)
-        if self.datas['form']['category']== 'Academics':
-            logo = None
-        elif self.datas['form']['category'] == 'Transport':
-            logo = company_recs[0].company_clogo_trans
+        for rec in company_recs:
+            logo = rec.company_clogo
         return logo
 
     def get_challan_header_lineone(self):
-        
         rescompany_id = self.pool.get('res.company').search(self.cr, self.uid,[])
         #-------------Handling Only one Company is There are multiple companies blank space will be returned----------------        
         if len(rescompany_id)>1:
             return ''
         company_recs = self.pool.get('res.company').browse(self.cr, self.uid, rescompany_id)
-        if self.datas['form']['category']== 'Academics':
-            line1 = company_recs[0].company_cfieldone
-        elif self.datas['form']['category'] == 'Transport':
-            line1 = company_recs[0].company_cfieldone_trans
-        return line1
+        for rec in company_recs:
+            fieldone = rec.company_cfieldone
+        return fieldone
 
     def get_challan_header_linetwo(self):
         rescompany_id = self.pool.get('res.company').search(self.cr, self.uid,[])
@@ -94,11 +73,9 @@ class unpaid_fee_challan_parser(report_sxw.rml_parse):
         if len(rescompany_id)>1:
             return ''
         company_recs = self.pool.get('res.company').browse(self.cr, self.uid, rescompany_id)
-        if self.datas['form']['category']== 'Academics':
-            line2 = company_recs[0].company_cfieldtwo
-        elif self.datas['form']['category'] == 'Transport':
-            lin2 = company_recs[0].company_cfieldtwo_trans
-        return line2
+        for rec in company_recs:
+            fieldtwo = rec.company_cfieldtwo
+        return fieldtwo
     
     def get_challan_header_linethree(self):
         rescompany_id = self.pool.get('res.company').search(self.cr, self.uid,[])
@@ -106,11 +83,9 @@ class unpaid_fee_challan_parser(report_sxw.rml_parse):
         if len(rescompany_id)>1:
             return ''
         company_recs = self.pool.get('res.company').browse(self.cr, self.uid, rescompany_id)
-        if self.datas['form']['category']== 'Academics':
-            line3 = company_recs[0].company_cfieldthree
-        elif self.datas['form']['category'] == 'Transport':
-            line3 = company_recs[0].company_cfieldthree_trans
-        return line3
+        for rec in company_recs:
+            fieldthree = rec.company_cfieldthree
+        return fieldthree
     
     def get_challan_footer_one(self):
         rescompany_id = self.pool.get('res.company').search(self.cr, self.uid,[])
@@ -118,11 +93,9 @@ class unpaid_fee_challan_parser(report_sxw.rml_parse):
         if len(rescompany_id)>1:
             return ''
         company_recs = self.pool.get('res.company').browse(self.cr, self.uid, rescompany_id)
-        if self.datas['form']['category']== 'Academics':
-            line4 = company_recs[0].company_cfieldfour
-        elif self.datas['form']['category'] == 'Transport':
-            line4 = company_recs[0].company_cfieldfour_trans
-        return line4
+        for rec in company_recs:
+            footerone = rec.company_cfieldfour
+        return footerone
     
     def get_challan_footer_two(self):
         rescompany_id = self.pool.get('res.company').search(self.cr, self.uid,[])
@@ -130,11 +103,9 @@ class unpaid_fee_challan_parser(report_sxw.rml_parse):
         if len(rescompany_id)>1:
             return ''
         company_recs = self.pool.get('res.company').browse(self.cr, self.uid, rescompany_id)
-        if self.datas['form']['category']== 'Academics':
-            line5 = company_recs[0].company_cfieldfive
-        elif self.datas['form']['category'] == 'Transport':
-            line5 = company_recs[0].company_cfieldfive_trans
-        return line5
+        for rec in company_recs:
+            footertwo = rec.company_cfieldfive
+        return footertwo
      
     def get_today(self):
         today = time.strftime('%d-%m-%Y')
@@ -154,66 +125,93 @@ class unpaid_fee_challan_parser(report_sxw.rml_parse):
         return group  
      
     def get_challans(self, data):
-        #currentlty this parser is set to call for whole class challans orinting
-        #both transport and acadmics
-        #later on this will be set to call this parser for single students also, both academics challans and transprot challans (3-10-2017)
+        
         
         challan_list = []
-        ###########print challan at the time of admission for paying fee (it is before admitting student)
+        print "this is unpaid_fee_challan_parser301........................."
         if data['model'] == 'student.admission.register':
             tlt_amount = 0
             rec = self.pool.get('student.admission.register').browse(self.cr,self.uid,data['active_ids'] )
             info_dict = {'name':rec.name,'father_name':rec.father_name,'Class':rec.student_class.name,'semester':''}
+            
             fee_res = []
-            challan_dict = {'challan_number':'','candidate_info':'','on_accounts':'','total_amount':'','amount_in_words':''}
+            challan_dict = {'banks_1':'','challan_number_1':'','candidate_info_1':'','on_accounts_1':'','total_amount_1':'','amount_in_words_1':'',
+                            'banks_2':'','challan_number_2':'','candidate_info_2':'','on_accounts_2':'','total_amount_2':'','amount_in_words_2':''}
+            i = 0
             for fee in rec.fee_ids  :
+                if (i % 2) == 0:
+                    index = 2
+                else:
+                    index = (i % 2)
+                print("index................  ",index)
+                    
                 fee_name = str(fee.name.name)+"  "+str(fee.fee_month.name)
                 dict = {'head_name':fee_name,'head_amount':fee.amount}
                 fee_res.append(dict)
                 tlt_amount = tlt_amount+fee.amount
-            challan_dict['challan_number'] = rec.registration_no
-            challan_dict['candidate_info'] =  [info_dict]
-            challan_dict['on_accounts'] = fee_res
-            challan_dict['total_amount'] = tlt_amount
+                
+                challan_dict['challan_number_'+ str(index)]  = rec.registration_no
+                challan_dict['candidate_info_'+ str(index)] =  [info_dict]
+                challan_dict['on_accounts_'+ str(index)] = fee_res
+                challan_dict['total_amount_'+ str(index)] = tlt_amount
+                
+                if i % 2 == 0:
+                    challan_list.append(challan_dict)
+                    challan_dict = {'banks_1':'','challan_number_1':'','candidate_info_1':'','on_accounts_1':'','total_amount_1':'','amount_in_words_1':'',
+                                    'banks_2':'','challan_number_2':'','candidate_info_2':'','on_accounts_2':'','total_amount_2':'','amount_in_words_2':''}
+                i = i + 1
+                
+            if (i-1) % 2 != 0:
+                challan_list.append(challan_dict)
+                
             #*******************convert the amount in text form****************************
             user_id = self.pool.get('res.users').browse(self.cr, self.uid,[self.uid])[0]
             cur = user_id.company_id.currency_id.name
             amt_en = amount_to_text_en.amount_to_text(tlt_amount,'en',cur);
             return_value=str(amt_en).replace('Cent','Paisa')
             #*******************************************************************************
+            
             challan_dict['amount_in_words'] = return_value
             challan_list.append(challan_dict)                     
         else:
-           #check if printed via student form
-            if 'student_id' in self.datas['form']:
-               #challan is being printed via student form, canlcel all other challans of this sutdent
-               challan_ids = self.pool.get('smsfee.receiptbook').search(self.cr, self.uid,[('challan_cat','=',self.datas['form']['category']),('student_id','=',self.datas['form']['student_id'][0]),('state','=','fee_calculated')])
-            else:
-               print "we are not printing via student form "
-               cls_id = self.datas['form']['class_id'][0]
-               challan_ids = self.pool.get('smsfee.receiptbook').search(self.cr, self.uid,[('challan_cat','=',self.datas['form']['category']),('student_class_id','=',cls_id),('state','=','fee_calculated')])
             
+            cls_id = self.datas['form']['class_id'][0]
+            challan_ids = self.pool.get('smsfee.receiptbook').search(self.cr, self.uid,[('student_class_id','=',cls_id),('state','=','fee_calculated')]) 
             if challan_ids:
-                rec_challan_ids = self.pool.get('smsfee.receiptbook').browse(self.cr, self.uid,challan_ids) 
+                i = 1
+                challan_dict = {'challan_number_1':'','candidate_info_1':'','vertical_lines_1':'','on_accounts_1':'','total_amount_1':'','amount_in_words_1':'','amount_after_due_date_1':'',
+                                'challan_number_2':'','candidate_info_2':'','vertical_lines_2':'','on_accounts_2':'','total_amount_2':'','amount_in_words_2':'','amount_after_due_date_2':''}
+                
+                rec_challan_ids = self.pool.get('smsfee.receiptbook').browse(self.cr, self.uid, challan_ids) 
                 for challan in rec_challan_ids:
-                    challan_dict = {'challan_number':'','candidate_info':'','on_accounts':'','vertical_lines':'','total_amount':'','amount_in_words':'','amount_after_due_date':'','dbid':''}
-                    challan_dict['challan_number'] = self.get_challan_number(challan.id)
-                    challan_dict['candidate_info'] = self.get_candidate_info(challan.student_id.id)
-                    challan_dict['on_accounts'] = self.get_on_accounts(challan.id)
-                    challan_dict['vertical_lines'] = self.get_vertical_lines_total(challan.id)
-                    challan_dict['total_amount'] = self.get_total_amount(challan.id)
-                    challan_dict['amount_in_words'] = self.get_amount_in_words(challan.id) 
-                    challan_dict['dbid'] = self.print_challan_dbid(challan.id) 
-                    if self.datas['form']['amount_after_due_date']:
-                        challan_dict['amount_after_due_date'] = challan_dict['total_amount'] + self.datas['form']['amount_after_due_date'] 
+                    if (i % 2) == 0:
+                        index = 2
                     else:
-                        challan_dict['amount_after_due_date'] = data['form']['amount_after_due_date']
+                        index = (i % 2)
+                    print("0000000000index................  ",index)
+                    challan_dict['challan_number_'+ str(index)] = self.get_challan_number(challan.id)
+                    challan_dict['candidate_info_'+ str(index)] = self.get_candidate_info(challan.student_id.id)
+                    challan_dict['on_accounts_'+ str(index)] = self.get_on_accounts(challan.id)
+                    challan_dict['vertical_lines_'+ str(index)] = self.get_vertical_lines_total(challan.id)
+                    challan_dict['total_amount_'+ str(index)] = self.get_total_amount(challan.id)
+                    challan_dict['amount_in_words_'+ str(index)] = self.get_amount_in_words(challan.id)
+                    if self.datas['form']['amount_after_due_date']:
+                        challan_dict['amount_after_due_date_'+str(index)] = challan_dict['total_amount_'+str(index)] + self.datas['form']['amount_after_due_date'] 
+                    else:
+                        challan_dict['amount_after_due_date_'+str(index)] = data['form']['amount_after_due_date']
+                    if i % 2 == 0:
+                        challan_list.append(challan_dict)
+                        challan_dict = {'challan_number_1':'','candidate_info_1':'','vertical_lines_1':'','on_accounts_1':'','total_amount_1':'','amount_in_words_1':'','amount_after_due_date_1':'',
+                                        'challan_number_2':'','candidate_info_2':'','vertical_lines_2':'','on_accounts_2':'','total_amount_2':'','amount_in_words_2':'','amount_after_due_date_2':''}
+                    i = i + 1
+
+                if (i-1) % 2 != 0:
                     challan_list.append(challan_dict)
         return challan_list  
      
     def get_user_name(self):
         user_name = self.pool.get('res.users').browse(self.cr, self.uid, self.uid, self.context).name
-        return   user_name
+        return user_name
  
     def get_vertical_lines(self, data):
         line_dots = []
@@ -238,8 +236,8 @@ class unpaid_fee_challan_parser(report_sxw.rml_parse):
         return line_dots    
      
     def get_challan_number(self, data):
-        challan_ids = self.pool.get('smsfee.receiptbook').search(self.cr, self.uid,[('id','=',data)])
-        challan_rec = self.pool.get('smsfee.receiptbook').browse(self.cr, self.uid,challan_ids)
+        challan_ids = self.pool.get('smsfee.receiptbook').search(self.cr, self.uid, [('id','=', data)])
+        challan_rec = self.pool.get('smsfee.receiptbook').browse(self.cr, self.uid, challan_ids)
         challan =  challan_rec[0].counter or str(challan_rec[0].id)+"*****"
         return challan
  
@@ -260,20 +258,20 @@ class unpaid_fee_challan_parser(report_sxw.rml_parse):
  
     def get_on_accounts(self, data):
         result = []
-        lines_ids = self.pool.get('smsfee.receiptbook.lines').search(self.cr, self.uid, [('receipt_book_id','=',data)])
+        lines_ids = self.pool.get('smsfee.receiptbook.lines').search(self.cr,self.uid, [('receipt_book_id','=',data)])
         if lines_ids:
-            challans = self.pool.get('smsfee.receiptbook.lines').browse(self.cr, self.uid, lines_ids)
+            challans = self.pool.get('smsfee.receiptbook.lines').browse(self.cr,self.uid,lines_ids)
         for challan in challans:
             dict = {'head_name':challan.fee_name,'head_amount':challan.fee_amount}
             result.append(dict) 
         return result
-    
-    def print_challan_dbid(self, data):
-        return data
      
     def get_total_amount(self, data):
-        receipt = self.pool.get('smsfee.receiptbook').browse(self.cr, self.uid, data)
-        total_amount_str = receipt.total_paybles
+        receipt = self.pool.get('smsfee.receiptbook').browse(self.cr,self.uid,data)
+        if self.datas['form']['amount_after_due_date']:
+            total_amount_str = receipt.total_paybles + self.datas['form']['amount_after_due_date'] 
+        else:
+            total_amount_str = receipt.total_paybles
         return total_amount_str
      
     def get_amount_in_words(self,data):
@@ -284,5 +282,4 @@ class unpaid_fee_challan_parser(report_sxw.rml_parse):
         return_value=str(amt_en).replace('Cent','Paisa')
         return return_value
      
-
-report_sxw.report_sxw('report.smsfee_print_one_student_per_page', 'smsfee.classfees.register', 'addons/smsfee/smsfee_unpaid_receipts_report_1o1.rml',parser = unpaid_fee_challan_parser, header=None)
+report_sxw.report_sxw('report.smsfee_print_three_student_per_page', 'smsfee.classfees.register', 'addons/smsfee/smsfee_unpaid_receipts_report_3o1.rml',parser = report_unpaid_fee_bills_2folded, header=None)
