+++ smsfee/smsfee_report.xml
@@ -90,7 +90,7 @@
             model="smsfee.classfees.register"
             name="smsfee_print_one_student_per_page"
             multi="True"
-            rml="smsfee/report/smsfee_unpaid_receipts_report_1o1.rml"
+            rml="smsfee/report/report_unpaid_fee_bills_3folded.rml"
             string="Fee Receipts (Unpaid)"
             />   
 
@@ -99,7 +99,7 @@
             model="smsfee.classfees.register"
             name="smsfee_print_three_student_per_page"
             multi="True"
-            rml="smsfee/report/smsfee_unpaid_receipts_report_3o1.rml"
+            rml="smsfee/report/report_unpaid_fee_bills_2folded.rml"
             string="Fee Receipts (Unpaid)"
             />
